# pkg/api/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AuthService"
    },
    {
      "name": "RankService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/add-phrases": {
      "post": {
        "operationId": "RankService_AddPhrases",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mainEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mainAddPhrasesReq"
            }
          }
        ],
        "tags": [
          "RankService"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "AuthService_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mainTokenMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mainUser"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/ranking": {
      "post": {
        "operationId": "RankService_Ranking",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mainRankingResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mainTokenMessage"
            }
          }
        ],
        "tags": [
          "RankService"
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "AuthService_RegisterUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mainTokenMessage"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mainUser"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    }
  },
  "definitions": {
    "mainAddPhrasesReq": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "phrases": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/mainPhrase"
          }
        }
      }
    },
    "mainEmpty": {
      "type": "object"
    },
    "mainKeyPhrase": {
      "type": "object",
      "properties": {
        "phrase": {
          "$ref": "#/definitions/mainPhrase"
        },
        "ranks": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/mainRank"
          }
        }
      }
    },
    "mainPhrase": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      }
    },
    "mainRank": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        },
        "rank": {
          "type": "integer",
          "format": "int32"
        },
        "paidRank": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "mainRankingResp": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "keyPhrases": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/mainKeyPhrase"
          }
        }
      }
    },
    "mainTokenMessage": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "mainUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}